+++ src/main/java/com/kalsym/order/service/model/repository/OrderRepository.java
@@ -3,9 +3,11 @@ package com.kalsym.order.service.model.repository;
 import com.kalsym.order.service.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 //import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+import java.util.List;
 
 /**
  *
@@ -17,4 +19,7 @@ public interface OrderRepository extends PagingAndSortingRepository<Order, Strin
 //    List<Order> findByCustomerId(@Param("customerId") String customerId);
     
 //    Page<Order> findAllByCreatedBetween(Date from, Date to, Pageable pageable);
+    
+    @Query("SELECT c.completionStatus, COUNT(c.id) FROM Order AS c GROUP BY c.completionStatus")
+    List<Object[]> getCountSummary();
 }