+++ src/main/java/com/kalsym/order/service/model/repository/StoreRepository.java
@@ -3,6 +3,8 @@ package com.kalsym.order.service.model.repository;
 import com.kalsym.order.service.model.Store;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -14,6 +16,13 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface StoreRepository extends PagingAndSortingRepository<Store, String>, JpaRepository<Store, String> {
-
-//    List<Order> findByCustomerId(@Param("customerId") String customerId);
+    
+    @Query(value = "SELECT getInvoiceSeqNo(:storeId)", nativeQuery = true)
+    int getInvoiceSeqNo(@Param("storeId") String storeId);
+    
+    @Modifying
+    @Query("UPDATE Store m SET m.invoiceSeqNo=0 WHERE m.id = :storeId") 
+    void ResetInvoiceSeqNo(
+            @Param("storeId") String storeId
+            );
 }