+++ src/main/java/com/kalsym/order/service/utility/TxIdUtil.java
@@ -3,6 +3,7 @@ package com.kalsym.order.service.utility;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Random;
+import com.kalsym.order.service.model.repository.StoreRepository;
 
 /**
  *
@@ -10,7 +11,7 @@ import java.util.Random;
  */
 public class TxIdUtil {
 
-    public static String generateReferenceId(String prefix) {
+    public static String generateReferenceIdOld(String prefix) {
         String referenceId = prefix;
         Date dNow = new Date();
         SimpleDateFormat ft = new SimpleDateFormat("yyMMddhhmmss");
@@ -21,6 +22,15 @@ public class TxIdUtil {
 
         referenceId = referenceId + datetime + n;
 
+        return referenceId;
+    }
+    
+    public static String generateInvoiceId(String storeId, String prefix, StoreRepository storeRepository) {
+        
+        //get running number for current store
+        int orderNo = storeRepository.getInvoiceSeqNo(storeId);
+        if (orderNo>=99999) { storeRepository.ResetInvoiceSeqNo(storeId); }
+        String referenceId = prefix + String.format("%05d", orderNo);
         return referenceId;
     }
 }